chore: move extractDomain to story.utils
